TRQ-2342: Merge fix from TRQ-2317/TRQ-1714: spool files left alone when MS is the same as pbs_server and there's an epilogue running.

+++ src/server/job_func.c
@@ -1813,7 +1813,8 @@ int record_jobinfo(
 
 int svr_job_purge(
 
-  job *pjob)  /* I (modified) */
+  job *pjob,  /* I (modified) */
+  int leaveSpoolFiles) /* I */
 
   {
   int           rc = PBSE_NONE;
@@ -1971,34 +1972,37 @@ int svr_job_purge(
       }
     }
 
-  /* delete any spooled stdout */
-  snprintf(namebuf, sizeof(namebuf), "%s%s%s", path_spool, job_fileprefix, JOB_STDOUT_SUFFIX);
-
-  if (unlink(namebuf) < 0)
+  if (!leaveSpoolFiles)
     {
-    if (errno != ENOENT)
-      log_err(errno, __func__, msg_err_purgejob);
-    }
-  else if (LOGLEVEL >= 6)
-    {
-    sprintf(log_buf, "removed job stdout");
+    /* delete any spooled stdout */
+    snprintf(namebuf, sizeof(namebuf), "%s%s%s", path_spool, job_fileprefix, JOB_STDOUT_SUFFIX);
 
-    log_record(PBSEVENT_DEBUG, PBS_EVENTCLASS_JOB, job_id, log_buf);
-    }
+    if (unlink(namebuf) < 0)
+      {
+      if (errno != ENOENT)
+        log_err(errno, __func__, msg_err_purgejob);
+      }
+    else if (LOGLEVEL >= 6)
+      {
+      sprintf(log_buf, "removed job stdout");
 
-  /* delete any spooled stderr ($TRQ_HOME/spool) */
-  snprintf(namebuf, sizeof(namebuf), "%s%s%s", path_spool, job_fileprefix, JOB_STDERR_SUFFIX);
+      log_record(PBSEVENT_DEBUG, PBS_EVENTCLASS_JOB, job_id, log_buf);
+      }
 
-  if (unlink(namebuf) < 0)
-    {
-    if (errno != ENOENT)
-      log_err(errno, __func__, msg_err_purgejob);
-    }
-  else if (LOGLEVEL >= 6)
-    {
-    sprintf(log_buf, "removed job stderr");
+    /* delete any spooled stderr ($TRQ_HOME/spool) */
+    snprintf(namebuf, sizeof(namebuf), "%s%s%s", path_spool, job_fileprefix, JOB_STDERR_SUFFIX);
 
-    log_record(PBSEVENT_DEBUG, PBS_EVENTCLASS_JOB, job_id, log_buf);
+    if (unlink(namebuf) < 0)
+      {
+      if (errno != ENOENT)
+        log_err(errno, __func__, msg_err_purgejob);
+      }
+    else if (LOGLEVEL >= 6)
+      {
+      sprintf(log_buf, "removed job stderr");
+
+      log_record(PBSEVENT_DEBUG, PBS_EVENTCLASS_JOB, job_id, log_buf);
+      }
     }
 
   /* delete checkpoint file directory if there is one */
